♻️(organizations) move organizations to the courses application
The Organization model is closely related to the Course model
so it does not make sense to have it in a separate application.
Reset the initial migration for all models in the courses
application because the project still has no running instance.
