+++ apps/courses/admin.py
@@ -5,7 +5,7 @@ from django.contrib import admin
 
 from cms.extensions import PageExtensionAdmin
 
-from .models import Course
+from .models import Course, Organization
 
 
 class CourseAdmin(PageExtensionAdmin):
@@ -32,4 +32,19 @@ class CourseAdmin(PageExtensionAdmin):
         form.instance.organizations.add(form.instance.main_organization)
 
 
+class OrganizationAdmin(PageExtensionAdmin):
+    """
+    Admin class for the Organization model
+    """
+    list_display = ["title", "code"]
+
+    # pylint: disable=no-self-use
+    def title(self, obj):
+        """
+        Get the page title from the related page
+        """
+        return obj.extended_object.get_title()
+
+
 admin.site.register(Course, CourseAdmin)
+admin.site.register(Organization, OrganizationAdmin)
