+++ apps/courses/cms_wizards.py
@@ -11,9 +11,7 @@ from cms.models import Page
 from cms.wizards.wizard_base import Wizard
 from cms.wizards.wizard_pool import wizard_pool
 
-from apps.organizations.models import Organization
-
-from .models import Course, Subject
+from .models import Course, Organization, Subject
 
 
 class BaseWizardForm(forms.Form):
@@ -165,11 +163,46 @@ wizard_pool.register(
 )
 
 
+class OrganizationWizardForm(BaseWizardForm):
+    """
+    This form is used by the wizard that creates a new organization page
+    A related organization model is created for each organization page
+    """
+
+    model = Organization
+
+    def save(self):
+        """
+        The parent form created the page.
+        This method creates the associated organization.
+        """
+        page = super().save()
+        Organization.objects.create(extended_object=page)
+        return page
+
+
+class OrganizationWizard(Wizard):
+    """Inherit from Wizard because each wizard must have its own Python class."""
+    pass
+
+
+wizard_pool.register(
+    OrganizationWizard(
+        title=_("New Organization page"),
+        description=_("Create a new Organization page"),
+        model=Organization,
+        form=OrganizationWizardForm,
+        weight=200,
+    )
+)
+
+
 class SubjectWizardForm(BaseWizardForm):
     """
     This form is used by the wizard that creates a new subject page.
     A related Subject model is created for each subject page.
     """
+
     model = Subject
 
     def save(self):
