+++ apps/courses/factories.py
@@ -1,13 +1,48 @@
 """
 Courses factories
 """
-import factory
+from django.utils.text import slugify
 
 from cms.api import create_page
+import factory
 
-from apps.organizations.factories import OrganizationFactory
+from .models import Course, Organization, Subject
 
-from .models import Course, Subject
+
+class OrganizationFactory(factory.django.DjangoModelFactory):
+    """
+    A factory to automatically generate random yet meaningful organization page extensions
+    in our tests.
+    """
+
+    class Meta:
+        model = Organization
+        exclude = ["title"]
+
+    logo = factory.django.ImageField(width=180, height=100)
+    title = factory.Faker("catch_phrase")
+
+    @factory.lazy_attribute
+    def extended_object(self):
+        """
+        Automatically create a related page with a random title
+        """
+        return create_page(self.title, "courses/cms/organization_detail.html", "en")
+
+    @factory.lazy_attribute
+    def code(self):
+        """
+        Since `name` is required, let's just slugify it to get a meaningful code (and keep it
+        below 100 characters)
+        """
+        return slugify(self.title)[:100]
+
+    @factory.post_generation
+    # pylint: disable=unused-argument, attribute-defined-outside-init, no-member
+    def with_courses(self, create, extracted, **kwargs):
+        """Add courses to ManyToMany relation."""
+        if create and extracted:
+            self.courses.set(extracted)
 
 
 class CourseFactory(factory.django.DjangoModelFactory):
