+++ apps/courses/tests/test_templates_organization_detail.py
